refresh dropdown when dependent field change
